+++ resources/js/callbacks.js
@@ -1,5 +1,5 @@
-Vue.prototype.scrollToId = (id) => {
-    let el = window.document.getElementById(id)
+Vue.prototype.scrollToElement = (selector) => {
+    let el = window.document.querySelector(selector)
     window.scrollTo({
         top: el.offsetTop,
         behavior: 'smooth',