+++ src/Http/ViewComposers/ConfigComposer.php
@@ -4,10 +4,10 @@ namespace Rapidez\Core\Http\ViewComposers;
 
 use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Cache;
+use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Str;
 use Illuminate\View\View;
 use Rapidez\Core\Facades\Rapidez;
-use Rapidez\Core\Models\Config;
 
 class ConfigComposer
 {
@@ -18,10 +18,10 @@ class ConfigComposer
             array_merge(config('rapidez.exposed'), ['store_code'])
         );
 
-        config(['frontend' => array_merge(
+        Config::set('frontend', array_merge(
             config('frontend') ?: [],
             $exposedFrontendConfigValues
-        )]);
+        ));
 
         $attributeModel = config('rapidez.models.attribute');
         $searchableAttributes = Arr::pluck(
@@ -30,18 +30,18 @@ class ConfigComposer
             'code'
         );
 
-        config(['frontend.locale' => Rapidez::config('general/locale/code', 'en_US')]);
-        config(['frontend.default_country' => Rapidez::config('general/country/default', 'NL')]);
-        config(['frontend.currency' => Rapidez::config('currency/options/default')]);
-        config(['frontend.cachekey' => Cache::rememberForever('cachekey', fn () => md5(Str::random()))]);
-        config(['frontend.redirect_cart' => (bool) Rapidez::config('checkout/cart/redirect_to_cart')]);
-        config(['frontend.show_swatches' => (bool) Rapidez::config('catalog/frontend/show_swatches_in_product_list')]);
-        config(['frontend.translations' => __('rapidez::frontend')]);
-        config(['frontend.recaptcha' => Config::getCachedByPath('recaptcha_frontend/type_recaptcha_v3/public_key', null, true)]);
-        config(['frontend.searchable' => array_merge($searchableAttributes, config('rapidez.searchable'))]);
-        config(['frontend.customer_fields_show' => $this->getCustomerFields()]);
-        config(['frontend.grid_per_page' => Rapidez::config('catalog/frontend/grid_per_page', 12)]);
-        config(['frontend.grid_per_page_values' => explode(',', Rapidez::config('catalog/frontend/grid_per_page_values', '12, 24'))]);
+        Config::set('frontend.locale', Rapidez::config('general/locale/code', 'en_US'));
+        Config::set('frontend.default_country', Rapidez::config('general/country/default', 'NL'));
+        Config::set('frontend.currency', Rapidez::config('currency/options/default'));
+        Config::set('frontend.cachekey', Cache::rememberForever('cachekey', fn () => md5(Str::random())));
+        Config::set('frontend.redirect_cart', (bool) Rapidez::config('checkout/cart/redirect_to_cart'));
+        Config::set('frontend.show_swatches', (bool) Rapidez::config('catalog/frontend/show_swatches_in_product_list'));
+        Config::set('frontend.translations', __('rapidez::frontend'));
+        Config::set('frontend.recaptcha', Rapidez::config('recaptcha_frontend/type_recaptcha_v3/public_key', null, true));
+        Config::set('frontend.searchable', array_merge($searchableAttributes, config('rapidez.searchable')));
+        Config::set('frontend.customer_fields_show', $this->getCustomerFields());
+        Config::set('frontend.grid_per_page', Rapidez::config('catalog/frontend/grid_per_page', 12));
+        Config::set('frontend.grid_per_page_values', explode(',', Rapidez::config('catalog/frontend/grid_per_page_values', '12, 24')));
     }
 
     public function getCustomerFields()