+++ src/RapidezServiceProvider.php
@@ -8,7 +8,12 @@ use Illuminate\Support\Facades\View;
 use Illuminate\Support\ServiceProvider;
 use Rapidez\Core\Commands\IndexProductsCommand;
 use Rapidez\Core\Commands\InstallCommand;
+use Rapidez\Core\Commands\InstallTestsCommand;
 use Rapidez\Core\Commands\ValidateCommand;
+use Rapidez\Core\Facades\Rapidez as RapidezFacade;
+use Rapidez\Core\Http\Controllers\Fallback\CmsPageController;
+use Rapidez\Core\Http\Controllers\Fallback\LegacyFallbackController;
+use Rapidez\Core\Http\Controllers\Fallback\UrlRewriteController;
 use Rapidez\Core\Http\Middleware\DetermineAndSetShop;
 use Rapidez\Core\Http\ViewComposers\ConfigComposer;
 use Rapidez\Core\ViewComponents\PlaceholderComponent;
@@ -47,6 +52,7 @@ class RapidezServiceProvider extends ServiceProvider
             IndexProductsCommand::class,
             ValidateCommand::class,
             InstallCommand::class,
+            InstallTestsCommand::class,
         ]);
 
         return $this;
@@ -78,12 +84,16 @@ class RapidezServiceProvider extends ServiceProvider
             $this->loadRoutesFrom(__DIR__.'/../routes/api.php');
         }
 
+        RapidezFacade::addFallbackRoute(UrlRewriteController::class, 5);
+        RapidezFacade::addFallbackRoute(CmsPageController::class, 10);
+        RapidezFacade::addFallbackRoute(LegacyFallbackController::class, 99999);
+
         return $this;
     }
 
     protected function bootThemes(): self
     {
-        $path = config('rapidez.themes.'.request()->server('MAGE_RUN_CODE', 'default'), false);
+        $path = config('rapidez.themes.'.request()->server('MAGE_RUN_CODE', request()->has('_store') && !app()->isProduction() ? request()->get('_store') : 'default'), false);
 
         if (!$path) {
             return $this;
@@ -153,7 +163,7 @@ class RapidezServiceProvider extends ServiceProvider
 
     protected function registerBindings(): self
     {
-        $this->app->bind('rapidez', Rapidez::class);
+        $this->app->singleton('rapidez', Rapidez::class);
         $this->app->bind('widget-directive', WidgetDirective::class);
 
         return $this;